Create Challenge1
In this first challenge one is supposed to create a program - with language of choice- that asks a user for his name, age, and account username. The information that is being asked was arbitrarily chosen, however, this demonstrates the countless applications of Java with  a program as simple as this.